refactor: extract PDF download logic out of reservationService
downloadPdf manipulates the DOM and browser APIs — not service layer
responsibility. Moved to src/utils/pdfUtils.js as downloadReservationPdf.
Removed diagnosePdfGeneration (dev-only diagnostic with no callers).
reservationSlice now imports directly from pdfUtils.
